+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.5.4] - TBA
+## [0.5.4] - 2021-12-05
 
 - #154 added Paystone, Ashby rules (thanks @austinhuang0131)
 - #146 added Liveclicker & eGain, improved Zeta Global, Customer.io, Granicus rules (thanks @m-schmitt)